requirements.txt
aniso8601==9.0.1
backcall==0.2.0
beautifulsoup4==4.9.3
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
python-dotenv==0.19.2
cymem==2.0.6
stripe==4.1.0
debugpy==1.5.1
decorator==5.1.0
Flask-Mail==0.9.1
dnspython==2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.3
pytesseract==0.3.9
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
sendgrid==6.9.7
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
lxml==4.8.0
gunicorn==20.1.0
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.0
jupyter-core==4.9.1
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
nest-asyncio==1.5.4
numpy==1.21.4
packaging==21.3
parso==0.8.3
pathy==0.6.1
pickleshare==0.7.5
preshed==3.0.6
prompt-toolkit==3.0.24
pydantic==1.8.2
Pygments==2.10.0
pymongo==4.0.1
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.7
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
